[IMP] Moved ac_api_return_access_denied into namespace blueprints
