+++ source/app/blueprints/pages/manage/manage_cases_routes.py
@@ -39,7 +39,7 @@ from app.iris_engine.access_control.utils import ac_current_user_has_permission
 from app.models.authorization import CaseAccessLevel
 from app.models.authorization import Permissions
 from app.schema.marshables import CaseDetailsSchema
-from app.util import ac_api_return_access_denied
+from app.blueprints.access_controls import ac_api_return_access_denied
 from app.blueprints.access_controls import ac_requires
 from app.util import response_error
