+++ source/app/blueprints/rest/case/case_assets_routes.py
@@ -59,8 +59,7 @@ from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
 from app.util import response_error
 from app.util import response_success
-from app.util import ac_api_return_access_denied
-
+from app.blueprints.access_controls import ac_api_return_access_denied
 
 case_assets_rest_blueprint = Blueprint('case_assets_rest', __name__)
