+++ source/app/blueprints/rest/manage/manage_assets_routes.py
@@ -25,7 +25,7 @@ from app.iris_engine.access_control.utils import ac_fast_check_current_user_has_
 from app.models.authorization import CaseAccessLevel
 from app.schema.marshables import CaseAssetsSchema
 from app.blueprints.access_controls import ac_api_requires
-from app.util import ac_api_return_access_denied
+from app.blueprints.access_controls import ac_api_return_access_denied
 from app.util import response_success
 
 manage_assets_rest_blueprint = Blueprint('manage_assets_rest', __name__)
