+++ source/app/blueprints/rest/manage/manage_cases_routes.py
@@ -52,7 +52,7 @@ from app.schema.marshables import CaseDetailsSchema
 from app.util import add_obj_history_entry
 from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
-from app.util import ac_api_return_access_denied
+from app.blueprints.access_controls import ac_api_return_access_denied
 from app.util import response_error
 from app.util import response_success
 from app.business.cases import cases_delete
